tests: Add qemujson2argvtest
This is basically a copy of the existing qemuxml2argvtest, with
a couple of obvious changes.

We're not using most of the features at the moment, but as the
JSON-based APIs evolve and get closer to feature-parity with
the XML-based ones we're definitely going to need more and more
of them, so we might as well start with the full package.