virtblocks: Save Virt Blocks language in a variable
We can then use it to remove lot of duplication from the rest of the code and
clean up the Makefiles.

Not completely, though, as on Ubuntu 18 (possibly related to old version of
autotools) using makefile variable in a filename breaks the build in what seems
to be an autotools bug.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>